+++ prebuild-generate-sitemap.js
@@ -69,6 +69,11 @@ async function generateSitemap() {
     }
   }
 
+  // Get article-pages URLs
+  if (config.component?.mainSideMenu?.items?.articles && config.articles?.length) {
+    urlCounter += generateArticleURLs(config.articles, urlOrigin, locale);
+  }
+
   // Get ebook URLs
   if (config.component?.mainSideMenu?.items?.ebooks && config.ebooks?.length) {
     urlCounter += generateEbookURLs(config.ebooks, urlOrigin, locale);
@@ -149,6 +154,16 @@ function generateAboutPagesURLs(aboutPages, mdFolderNumber, urlOrigin, locale) {
       appendToSitemapFile(url + '\n');
       counter += 1;
     }
+  }
+  return counter;
+}
+
+function generateArticleURLs(articles, urlOrigin, locale) {
+  let counter = 0;
+  for (let i = 0; i < articles.length; i++) {
+    const url = `${urlOrigin}/${locale}/article/${articles[i]['routeName']}`;
+    appendToSitemapFile(url + '\n');
+    counter += 1;
   }
   return counter;
 }