+++ src/assets/config/config.ts
@@ -34,6 +34,7 @@ export const config: Config = {
       collectionSideMenu: false
     }
   },
+  articles: [],
   collections: {
     addTEIClassNames: false,
     replaceImageAssetsPaths: false,
@@ -63,6 +64,10 @@ export const config: Config = {
   page: {
     about: {
       initialPageNode: "01"
+    },
+    article: {
+      showTextDownloadButton: false,
+      showURNButton: false
     },
     elasticSearch: {
       enableFilters: true,
@@ -267,6 +272,7 @@ export const config: Config = {
       categoricalSortingSecondaryKey: ""
     },
     contentGrid: {
+      includeArticles: false,
       includeEbooks: false,
       includeMediaCollection: false,
       mediaCollectionCoverURL: "",
@@ -282,6 +288,7 @@ export const config: Config = {
     mainSideMenu: {
       items: {
         about: true,
+        articles: false,
         ebooks: false,
         collections: true,
         mediaCollections: false,
@@ -291,7 +298,8 @@ export const config: Config = {
         indexWorks: false,
         search: true
       },
-      defaultExpanded: false
+      defaultExpanded: false,
+      ungroupArticles: false
     },
     manuscripts: {
       showTitle: true,